Created login view and cleaned up Program.cs
